backend/requirements.txt
fastapi==0.115.6
pydantic-settings==2.7.0
transformers==4.46.3
torch==2.5.1
uvicorn[standard]==0.32.1
